preprocess/pyproject.toml
[project]
name = "preprocess"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow==3.0.3",
    "pandas>=2.3.1",
    "pendulum>=3.1.0",
]
